mini-amazon-skeleton-main/pyproject.toml
[tool.poetry]
name = "mini-amazon"
version = "1.2023.0"
authors = ["CompSci 316 Teaching Staff <compsci-316@duke.edu>"]
description = "Mini Amazon Course Project"

[tool.poetry.dependencies]
python = "^3.11"
flask = "2.3.3"
werkzeug = "2.3.7"
flask-sqlalchemy = "^3.0.5"
flask-wtf = "^1.1.1"
psycopg2-binary = "^2.9.7"
flask-login = "^0.6.2"
email-validator = "^2.0.0.post2"
faker = "^19.3.1"
python-dotenv = "^1.0.0"
better-profanity = "^0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
